% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTarCAObject.R
\name{CreateTarCAObject}
\alias{CreateTarCAObject}
\title{Create a TarCA object}
\usage{
CreateTarCAObject(Tree, Ann, FileOut = NULL)
}
\arguments{
\item{Tree}{Tree object from read.tree.}

\item{Ann}{Dataframe with columns TipLabel and TipAnn.}

\item{FileOut}{Path to saving the TarCA object. (default: NULL)}
}
\value{
A TarCA object.
}
\description{
Creating a TarCA object with the tree object and the annotation dataset of tips
}
\examples{
Data.TarCA <- CreateTarCAObject(Var.Tree,Var.Ann)
}
