\name{example_data}
\alias{example_data}
\docType{data}
\title{
Numeric data frame of example data
}
\description{
Contains 20 genes across 564 single-cell sequencing experiments, sorted by pseudotime using the package Monocle.
}
\usage{data("example_data")}
\format{
  A data frame with 20 observations on 564 variables.
}
\details{
See vignette for more details.
}
\source{

Shalek AK, Satija R., Shuga J., Trombetta J.J. et al. (2014) Single-cell RNA-seq reveals dynamic paracrine control of cellular variation. Nature, 510(7505), 363-369. http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4193940/
}
\examples{
data(example_data)
}
\keyword{datasets}
