% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.AssignLaterState.R
\name{fun.AssignLaterState}
\alias{fun.AssignLaterState}
\title{Title}
\usage{
fun.AssignLaterState(TarCAObject, Predictions)
}
\arguments{
\item{TarCAObject}{A TarCA object obtained from CreateTarCAObject.}

\item{Predictions}{A dataframe of transferred result.}
}
\value{
A dataframe with later state.
}
\description{
Title
}
\examples{
fun.AssignLaterStage(Data.TarCA,Var.Predictions)
}
