% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.CalculatePotency.R
\name{fun.CalculatePotency}
\alias{fun.CalculatePotency}
\title{Calculating potency for each internal node..}
\usage{
fun.CalculatePotency(
  TarCAObject,
  Nodes,
  Threads = 4,
  DirOut = "LEAPOutput/",
  SaveObject = TRUE
)
}
\arguments{
\item{TarCAObject}{A TarCA object obtained from CreateTarCAObject.}

\item{Nodes}{A vector of internal nodes.}

\item{Threads}{Number of cores. (default: 4)}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{SaveObject}{Saving all objects. (default: TRUE)}
}
\value{
A dataframe with potency information for each internal node.
}
\description{
Mapping inferred state to the nodes and their unsolved descendants.
}
\examples{
Var.Potency <- CalculatePotency(Data.TarCA,Var.PredictedState$NodeLabel)
}
