% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.DetLUG.R
\name{fun.DetLUG}
\alias{fun.DetLUG}
\title{Title}
\usage{
fun.DetLUG(ObsLUG, ShufLUG, p_threshold = 0.01, Threads = 4)
}
\arguments{
\item{ObsLUG}{A dataframe of observing LUGs obtained from fun.LoadInLUGData.}

\item{ShufLUG}{A dataframe of shuffling LUGs obtained from fun.RunShufLUGEstimator.}

\item{p_threshold}{Threshold of empirical P-value. (default: 0.01)}

\item{Threads}{Number of cores. (default: 4)}
}
\value{
A dataframe of inferred LUGs.
}
\description{
Title
}
\examples{
Var.LUGs <- fun.DetLUG(Var.Obs,Var.Shuf)
}
