% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.ExtractLUG.R
\name{fun.ExtractLUG}
\alias{fun.ExtractLUG}
\title{Extract the LUGs from cell phylogeny and corresponding transcritptome.}
\usage{
fun.ExtractLUG(
  TarCAObject,
  ExpreMatrix,
  DirOut = "LEAPOutput/",
  TopGene = 0.1,
  NumShuf = 1000,
  p_threshold = 0.01,
  SeedFrom = NULL,
  Threads = 4,
  Redo = FALSE
)
}
\arguments{
\item{TarCAObject}{A TarCA object obtained from CreateTarCAObject.}

\item{ExpreMatrix}{A matrix of expression levels corresponding to the cell phylogeny.}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{TopGene}{Threshold of highly expressed population. (default: 0.1)}

\item{NumShuf}{Number of shuffling. (default: 1000)}

\item{p_threshold}{Threshold of empirical P-value. (default: 0.01)}

\item{SeedFrom}{Set the beginning of seed. (default: NULL)}

\item{Threads}{Number of cores. (default: 4)}

\item{Redo}{Redo LUGs searching. (default: FALSE)}
}
\value{
A dataframe of inferred LUGs.
}
\description{
LUGs inference based on permutation test.
}
\examples{
Data.TarCA <- CreateTarCAObject(Var.Tree,Var.Ann)
Var.LUGs <- fun.ExtractLUG(Data.TarCA,Var.TPM)
}
