% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.GetInternalState.R
\name{fun.GetInternalState}
\alias{fun.GetInternalState}
\title{Title}
\usage{
fun.GetInternalState(PhyloDepth, MinAct = 3, Threads = 4)
}
\arguments{
\item{PhyloDepth}{A dataframe of PhyloDepth obtained from fun.GetPhyloDepthOfLUGs.}

\item{MinAct}{Minimum number of activation events for reconstructed internal nodes. (default: 3)}

\item{Threads}{Number of cores. (default: 4)}
}
\value{
A matrix of inferred transcriptome of internal nodes.
}
\description{
Title
}
\examples{
fun.GetInternalState(Var.PhyloDepth)
}
