% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.GetPhyloDepthOfLUGs.R
\name{fun.GetPhyloDepthOfLUGs}
\alias{fun.GetPhyloDepthOfLUGs}
\title{Title}
\usage{
fun.GetPhyloDepthOfLUGs(TarCAObject, LUG, DirOut = "LEAPOutput/", Threads = 4)
}
\arguments{
\item{TarCAObject}{A TarCA object obtained from CreateTarCAObject.}

\item{LUG}{A dataframe of LUGs obtained from fun.ExtractLUG.}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{Threads}{Number of cores. (default: 4)}
}
\value{
A dataframe of normalized depth of internal nodes with LUG activation events.
}
\description{
Title
}
\examples{
fun.GetPhyloDepthOfLUGs(Data.TarCA,Var.LUGs)
}
