% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.LoadInLUGData.R
\name{fun.LoadInLUGData}
\alias{fun.LoadInLUGData}
\title{Title}
\usage{
fun.LoadInLUGData(DirOut = "LEAPOutput/", Threads = 4)
}
\arguments{
\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{Threads}{Number of cores. (default: 4)}
}
\value{
A dataframe of observed LUGs from local files.
}
\description{
Title
}
\examples{
Var.Obs <- fun.LoadInLUGData()
}
