% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.message.R
\name{fun.message}
\alias{fun.message}
\title{Title}
\usage{
fun.message(...)
}
\arguments{
\item{...}{Message you want to print.}
}
\value{
Message with time.
}
\description{
Title
}
\examples{
fun.message("Hello.")
}
