% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.ProcessSeurat.R
\name{fun.ProcessSeurat}
\alias{fun.ProcessSeurat}
\title{Title}
\usage{
fun.ProcessSeurat(SeuratObject, nPC = 30, Verbose = FALSE)
}
\arguments{
\item{SeuratObject}{A Seurat object.}

\item{nPC}{Number of PC.}

\item{Verbose}{Verbose or not.}
}
\value{
A processed Seurat object
}
\description{
Title
}
\examples{
fun.ProcessSeurat(Var.QuerySeurat)
}
