% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.RunLUGEstimator.R
\name{fun.RunLUGEstimator}
\alias{fun.RunLUGEstimator}
\title{Title}
\usage{
fun.RunLUGEstimator(
  TarCAObject,
  ExpreMatrix,
  DirOut = "LEAPOutput/",
  TopGene = 0.1,
  Threads = 4,
  Redo = FALSE
)
}
\arguments{
\item{TarCAObject}{A TarCA object obtained from CreateTarCAObject.}

\item{ExpreMatrix}{A matrix of expression levels corresponding to the cell phylogeny.}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{TopGene}{Threshold of highly expressed population. (default: 0.1)}

\item{Threads}{Number of cores. (default: 4)}

\item{Redo}{Redo LUGs searching. (default: FALSE)}
}
\value{
NULL.
}
\description{
Title
}
\examples{
fun.RunLUEEstimator(Data.TarCA,Var.TPM)
}
