% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.RunShufLUGEstimator.R
\name{fun.RunShufLUGEstimator}
\alias{fun.RunShufLUGEstimator}
\title{Title}
\usage{
fun.RunShufLUGEstimator(
  TarCAObject,
  ObsLUG,
  DirOut = "LEAPOutput/",
  NumShuf = 1000,
  SeedFrom = NULL,
  Threads = 4,
  Redo = FALSE
)
}
\arguments{
\item{ObsLUG}{A dataframe of LUGs obtained from fun.LoadInLUGData.}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{NumShuf}{Number of shuffling. (default: 1000)}

\item{SeedFrom}{Set the beginning of seed. (default: NULL)}

\item{Threads}{Number of cores. (default: 4)}

\item{Redo}{Redo LUGs searching. (default: FALSE)}

\item{TarCAObjeLUG}{A TarCA object obtained from CreateTarCAObject.}
}
\value{
A dataframe of shuffling LUGs.
}
\description{
Title
}
\examples{
Var.Shuf <- fun.RunShufLUGEstimator(Data.TarCA,Var.Obs)
}
