% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInternalMatrix.R
\name{GetInternalMatrix}
\alias{GetInternalMatrix}
\title{Reconstruct the LUG-transcriptome of internal nodes.}
\usage{
GetInternalMatrix(
  TarCAObject = NULL,
  ExpreMatrix = NULL,
  LUGs = NULL,
  PhyloDepth = NULL,
  DirOut = "LEAPOutput/",
  TopGene = 0.1,
  NumShuf = 1000,
  p_threshold = 0.01,
  MinAct = 3,
  SeedFrom = NULL,
  Threads = 4,
  SaveObject = TRUE,
  Redo = FALSE
)
}
\arguments{
\item{TarCAObject}{A TarCA object obtained from CreateTarCAObject. (default: NULL)}

\item{ExpreMatrix}{A matrix of expression levels corresponding to the cell phylogeny. (default: NULL)}

\item{LUGs}{A dataframe of LUGs obtained from fun.ExtractLUG. (optional,default: NULL)}

\item{PhyloDepth}{A dataframe of PhyloDepth obtained from fun.GetPhyloDepthOfLUGs. (optional,default: NULL)}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{TopGene}{Threshold of highly expressed population. (default: 0.1)}

\item{NumShuf}{Number of shuffling. (default: 1000)}

\item{p_threshold}{Threshold of empirical P-value. (default: 0.01)}

\item{MinAct}{Minimum number of activation events for reconstructed internal nodes. (default: 3)}

\item{SeedFrom}{Set the beginning of seed. (default: NULL)}

\item{Threads}{Number of cores. (default: 4)}

\item{SaveObject}{Saving all objects. (default: TRUE)}

\item{Redo}{Redo LUGs searching. (default: FALSE)}
}
\value{
A matrix of inferred transcriptome of internal nodes.
}
\description{
Reconstruct the transcriptome of internal nodes based on LUG inference.
}
\examples{
Data.TarCA <- CreateTarCAObject(Var.Tree,Var.Ann)
Var.InternalState <- GetInternalMatrix(Data.TarCA,Var.TPM)
}
