% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredictedState.R
\name{GetPredictedState}
\alias{GetPredictedState}
\title{Get Inferred states from label transfer data.}
\usage{
GetPredictedState(TarCAObject, Predictions, RemoveRoot = NULL)
}
\arguments{
\item{TarCAObject}{A TarCA object obtained from CreateTarCAObject.}

\item{Predictions}{A dataframe of transferred result.}

\item{RemoveRoot}{Remove root's name.}
}
\value{
A datafraome of impuated annotation on the cell phylogeny.
}
\description{
Mapping inferred state to the nodes and their unsolved descendants.
}
\examples{
GetPredictedState(Data.TarCA,Var.Predictions)
}
