% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReferenceMatrix.R
\name{GetReferenceMatrix}
\alias{GetReferenceMatrix}
\title{Binarize reference dataset.}
\usage{
GetReferenceMatrix(
  ExpreMatrix,
  Metadata,
  GroupBy = NULL,
  TopGene = 0.1,
  Threads = 4,
  DirOut = "LEAPOutput/",
  SaveObject = TRUE
)
}
\arguments{
\item{ExpreMatrix}{A matrix of expression levels corresponding to the reference atlas.}

\item{Metadata}{A dataframe containing metadata for each reference cell.}

\item{GroupBy}{Column name used to group reference data. (default: NULL)}

\item{TopGene}{Threshold of highly expressed population. (default: 0.1)}

\item{Threads}{Number of cores. (default: 4)}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{SaveObject}{Saving all objects. (default: TRUE)}
}
\value{
A binarized matrix of reference data.
}
\description{
Convert reference dataset into binarized matrix.
}
\examples{
GetReferenceMatrix(Data.Ref_Count,Data.Ref_metadata,GroupBy="embryo.time.bin")
}
