\name{lag_example}
\alias{lag_example}
\docType{data}
\title{
Integer data frame
}
\description{
Contains the resulting lag matrix from applying MAC_counter() to example_data
}
\usage{data("lag_example")}
\format{
  A data frame with 20 observations on 20 variables.
}
\details{
See vignette for more details.
}
\examples{
data(lag_example)
}
\keyword{datasets}
