% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEAP-package.R
\docType{package}
\name{LEAP-package}
\alias{LEAP}
\alias{LEAP-package}
\title{LEAP: LUG-Encoded Ancestral Projection}
\description{
A phylogeny-based computational framework for inferring the transcriptional
states of internal nodes from cell phylogenies via lineage-specific
upregulated genes (LUGs).
}
\author{
\strong{Maintainer}: Shanjun Deng \email{dengshj8@mail.sysu.edu.cn} [funder]

Authors:
\itemize{
  \item Shanjun Deng \email{dengshj8@mail.sysu.edu.cn} [funder]
  \item Xionglei He [funder]
  \item Zhan Liu [contributor]
}

}
