\name{leap}
\alias{leap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Leap
}
\description{
It calculates whether a year is a leap year or not
}
\usage{
leap(year)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{year}{
numeric. year 
}
}
\details{
none
}
\value{
logical
}
\references{
wikipedia
}
\author{
Loic
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
leap(2000)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ leap }
\keyword{ wikipedia }% __ONLY ONE__ keyword per line
