\name{MAC_example}
\alias{MAC_example}
\docType{data}
\title{
Numerical matrix
}
\description{
Contains the matrix of MACs that result from applying MAC_counter() to example_data}
\usage{data("MAC_example")}
\format{
  The format is:
 num [1:20, 1:20] NA 0.141 0.418 0.253 0.164 ...
}
\details{
See vignette for more details.
}
\examples{
data(MAC_example)
}
\keyword{datasets}
