\name{MAC_symmetric}
\alias{MAC_symmetric}
\docType{data}
\title{
Numeric data frame
}
\description{
Contains the results of applying MAC_counter(symmetric=T) to example_data
}
\usage{data("MAC_symmetric")}
\format{
  A data frame with 20 observations on 20 variables.
}
\examples{
data(MAC_symmetric)
}
\keyword{datasets}
