\name{perm_example}
\alias{perm_example}
\docType{data}
\title{
The resulting data ouptut from applying MAC_perm() to example_data
}
\description{
101 observations at correlation cutoffs, from 0 to 1, of the four variables Cors, MACs_observed, MACs_ave_perm, and fdr.
}
\usage{data("perm_example")}
\format{
  A data frame with 101 observations on 4 variables.
}
\details{
See vignette for more details
}
\examples{
data(perm_example)
}
\keyword{datasets}
