% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\docType{data}
\name{simulation_data}
\alias{simulation_data}
\title{Simulated treatment episode data}
\format{
An object of class \code{list} of length 2.
}
\usage{
simulation_data
}
\description{
Simulated longitudinal data for demonstrating treatment episode analyses.
}
\examples{
data(simulation_data)
head(simulation_data)
}
\keyword{datasets}
