% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_episode_count.R
\name{add_episode_count}
\alias{add_episode_count}
\title{Classify clients by treatment attendance}
\usage{
add_episode_count(data)
}
\arguments{
\item{data}{A data frame containing psychotherapy session records.}

\item{client}{Character string specifying the name of the client identifier
variable.}

\item{episode}{Character string specifying the name of the treatment
episode identifier variable.}
}
\value{
A data frame with two additional variables:
\itemize{
\item \code{n_episodes}
}
}
\description{
Creates client-level attendance variables based on the number of treatment
episodes attended.This variable can easily be used to subset clients
based based on the total number of episodes they attended.
}
\details{
One variable is added to the data:

\describe{
\item{n_episodes}{Total number of treatment episodes attended by the client.}
}

This function assumes that treatment episodes have already been identified
using \code{\link[=add_episode_id]{add_episode_id()}}
}
\examples{
# Create attendance classification variables
df <- add_episode_count(data = df, client = "client_id", episode = "episode_id")

# with native R pipe
add_episode_session(df, client = "client_id", episode = "episode_id") |>
    add_episode_count("client_id", "episode_id")

# with magrittr pipe
add_episode_session(df, client = "client_id", episode = "episode_id") \%>\%
    add_episode_count("client_id", "episode_id")

# Select clients who attended multiple treatment episodes (n_episodes > 1)
subset(df, n_episodes > 1)

# Select clients who attended exactly three treatment episodes
subset(df, n_episodes == 3)

# Select clients who attended three or more treatment episodes
subset(df, n_episodes >= 3)

}
