% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_episode_id.R
\name{add_episode_id}
\alias{add_episode_id}
\title{Identify unique treatment episodes}
\usage{
add_episode_id(data, delimiter = 90)
}
\arguments{
\item{data}{A longitudinal data frame}

\item{client}{unique client identifier}

\item{session_lag}{lag time between sessions}

\item{threshold}{elapsed time used to demarcate treatment episodes}
}
\value{
A data frame with an additional variable, \code{episode_id},
indicating the episode membership of each session.
}
\description{
Assigns a unique episode number to each session record within a client.
A new treatment episode is identified whenever the elapsed time between
consecutive sessions exceeds a user-defined threshold. By default,
treatment episodes are separated by 90 or more days, but this default
value can be altered by the user by specifying a different value
for the threshold argument.
}
\details{
This function assumes records are sorted chronologically within client and
that session lags have already been calculated. Users should first apply
\code{\link[=order_sessions]{order_sessions()}} followed by \code{\link[=add_session_lag]{add_session_lag()}} before identifying
treatment episodes.
}
\examples{
out <- add_episode_id(data = df, client = "client_id", session_lag = "lag")

with native R pipe
add_session_lag("client_id", "session_date") |>
    add_episode_id("client_id", "session_lag")

# with miggrtir pipe
add_session_lag("client_id", "session_date") \%>\%
    add_episode_id("client_id", "session_lag")

}
