% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_episode_session.R
\name{add_episode_session}
\alias{add_episode_session}
\title{Add session number within treatment episode}
\usage{
add_episode_session(data, client, episode)
}
\arguments{
\item{data}{A data frame containing session-level psychotherapy records.}

\item{client}{Character string specifying the name of the client identifier variable.}

\item{episode}{Character string specifying the name of the treatment episode identifier variable.}

\item{...}{Additional arguments. Currently unused.}
}
\value{
A data frame with an additional variable, \code{episode_session}, indicating
the consecutive session number within each treatment episode.
}
\description{
Creates a consecutive session number within each client and treatment
episode. The resulting variable indicates the order of each session within
its episode.
}
\details{
This function assumes that records are already sorted chronologically within
client and episode. Use \code{\link[=order_sessions]{order_sessions()}} before identifying episodes and
adding episode-specific session numbers.
}
\examples{
out <- add_episode_session(data = dat_unb, client = "client_id", episode = "episode_id")

# native R pipe
add_session_lag(data = dat_unb, client = "client_id", "session_date") |>
    add_episode_id(client = "client_id", session = "session_lag") |>
    add_episode_session(client = "client_id", episode = "episode_id")

# magrittr pipe
dat_unb \%>\%
   add_session_lag(client_id, session_date) \%>\%
   add_episode_id(client_id, session_lag) \%>\%
   add_episode_session(client_id, episode_id)

}
