\name{LEAP-package}
\alias{LEAP-package}
\alias{LEAP}
\docType{package}
\title{
\packageTitle{LEAP}
}
\description{
\packageDescription{LEAP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LEAP}
\packageIndices{LEAP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LEAP}

Maintainer: \packageMaintainer{LEAP}
}
\references{
Shalek AK, Satija R., Shuga J., Trombetta J.J. et al. (2014) Single-cell RNA-seq reveals dynamic paracrine control of cellular variation. Nature, 510(7505), 363-369. http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4193940/
}
\keyword{ package }
\examples{
\dontrun{MAC_results = MAC_counter(data=example_data)}

\dontrun{MAC_perm(data=example_data, MACs_observ=MAC_example)}
}
