\name{leap-package}
\alias{leap-package}
\alias{leap}
\docType{package}
\title{
Package to calculate leap year
}
\description{
Calculate a leap year
}
\details{
\tabular{ll}{
Package: \tab leap\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-08\cr
License: \tab wur\cr
}
leap is the most important function
}
\author{
Loic
}
\references{
Wikipedia
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ leap }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
leap(2000)
}
