% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAC_counter.R
\name{MAC_counter}
\alias{MAC_counter}
\title{compute}
\usage{
MAC_counter(
  data,
  max_lag_prop = 1/3,
  MAC_cutoff = 0.2,
  file_name = FALSE,
  lag_matrix = TRUE,
  symmetric = FALSE
)
}
\arguments{
\item{data}{data frame of gene expression data}

\item{max_lag_prop}{maximum lag to consider}

\item{MAC_cutoff}{minimum MAC value to consider}

\item{file_name}{name of file to write results to}

\item{lag_matrix}{whether to write lag matrix to file}

\item{symmetric}{whether to compute MACs for both symmetric and asymmetric lags}
}
\value{
MACs and lags
}
\description{
compute
}
\examples{
library(LEAP)
data(example_data)
results <- MAC_counter(
  data = example_data,
  max_lag_prop = 1 / 3,
  MAC_cutoff = 0.2,
  lag_matrix = T
)
}
