% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAC_lags.R
\name{MAC_lags}
\alias{MAC_lags}
\title{Compute MACs and lags}
\usage{
MAC_lags(data, max_lag_prop = 1/3, symmetric = FALSE)
}
\arguments{
\item{data}{data frame of gene expression data}

\item{max_lag_prop}{maximum lag to consider}

\item{symmetric}{whether to compute MACs for both symmetric and asymmetric lags}
}
\value{
MACs and lags
}
\description{
Compute MACs and lags
}
