% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAC_perm.R
\name{MAC_perm}
\alias{MAC_perm}
\title{permute}
\usage{
MAC_perm(
  data,
  MACs_observ,
  num_perms = 100,
  max_lag_prop = 1/3,
  FDR_cutoffs = 101,
  perm_file_name = F
)
}
\arguments{
\item{data}{data frame of gene expression data}

\item{MACs_observ}{MACs from observed data}

\item{num_perms}{number of permutations to run}

\item{max_lag_prop}{maximum lag to consider}

\item{FDR_cutoffs}{number of FDR cutoffs to calculate}

\item{perm_file_name}{name of file to write results tos}
}
\value{
permuted MACs and lags
}
\description{
permute
}
