% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunLabelTransfer.R
\name{RunLabelTransfer}
\alias{RunLabelTransfer}
\title{Run label transfer from reference data to query data.}
\usage{
RunLabelTransfer(
  QueryMatrix = NULL,
  RefMatrix = NULL,
  TransferLabel,
  QuerySeurat = NULL,
  RefSeurat = NULL,
  Anchors = NULL,
  nPC = 30,
  Verbose = FALSE,
  DirOut = "LEAPOutput/",
  SaveObject = TRUE
)
}
\arguments{
\item{QueryMatrix}{A binarized matrix of query data obtained from GetInternalMatrix.}

\item{RefMatrix}{A binarized matrix of reference data obtained from GetReferenceMatrix.}

\item{TransferLabel}{Transfer Labels from reference to query.}

\item{QuerySeurat}{Existed QuerySeurat.}

\item{RefSeurat}{Existed RefSeurat.}

\item{Anchors}{Existed Anchor.}

\item{nPC}{Number of PC. (default: 30)}

\item{Verbose}{Verbose or not. (default: FALSE)}

\item{DirOut}{Path to saving the LUG object. (default: "LEAPOutput/")}

\item{SaveObject}{Saving all objects. (default: TRUE)}
}
\value{
A dataframe of transferred result.
}
\description{
Run label transfer from reference data to query data, both of which were binarized by LUG.
}
\examples{
RunLabelTransfer(Var.InternalMatrix,Var.ReferenceMatrix,Data.Ref_metadata$lineage)
}
